Исправение поля в result.go

+++ backend/internal/models/result.go
@@ -4,9 +4,9 @@ import "gorm.io/gorm"
 
 type Result struct {
 	gorm.Model
-	UserID    uint `gorm:"index" json:"user_id"`
-	QuizID    uint `gorm:"index" json:"quiz_id"`
-	Score     int  `json:"score"`
-	MaxScore  int  `json:"max_score"`
-	TimeTaken int  `json:"time_taken"`
+	UserID    uint    `gorm:"index" json:"user_id"`
+	QuizID    uint    `gorm:"index" json:"quiz_id"`
+	Score     float64 `json:"score"`
+	MaxScore  int     `json:"max_score"`
+	TimeTaken int     `json:"time_taken"`
 }
